Several fixes to our use of splice
The splice manpage explicitly requires loff_t as offset. Copy "offset" there.
Probably not required, but I wanted to make sure it's as required.

Splice blocks with large buffers. For me it worked with 16k and blocked with
32k and beyond. It would be nice to see a clarification in the manpage of
splice for this behaviour.

Splice if used with an offset increments the offset. From the manpage this was
at least not entirely obvious :-)

I haven't yet activated this (try_splice_call ist still false by default), it
needs more testing.

Volker