s3: Remove a few unnecessary checks from the streams xattr module
Handling of error codes when renaming a file to a stream and a stream
to a file is now done in rename_internals_fsp.

The NTRENAME stream path only passes in the stream name, so the new
base can now be different from the old base.
(cherry picked from commit 078c4fbab8a5be2e23acda1978fd6d6071898dd9)