mount.cifs: initialize rc to 0 in main
The value of rc in main() isn't initialized in the declaration. This
wasn't a problem before, but Shirish's fakemount patch can make it so
that we return the uninitialized variable if the -n flag is used.

Fix this by initializing rc to 0.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 45d75ece3e35b5180e82dc0488b13fbd2a93c41e)