Add another torture test inspired by George @ Apple. Inside a directory, keep a file open and then renaming the directory should fail with ACCESS_DENIED. This is connected to the test case where the close was failing due to a delayed write on a file not being able to succeed when Samba allowed the containing directory to be renamed.
I will fix this in the server shortly (this should be done
across connections also but with will be very hard
in Samba - would need a full scan of the open file
db on every directory rename) - so I will fix for
the local case first (scanning local file opens
inside an smbd is cheap).

Jeremy.