s4:ldb Fix the paged_searches module
This simplifies the code, removes presumptions about being the only
control in the reply, and allows it to function against Windows 2008.

For searches which did not require a paged result, the module was
simply returning a failure when the compleated search did not include
a paged result control.

Andrew Bartlett