Use the server config file for smbpasswd -a user in make test
Without this, smbpasswd generated a SID for the netbios name TORTURE_6. This SID
somehow ended up in the user's token for quite a number of tests, leading to
NT_STATUS_ACCESS_DENIED. I haven't fully understood what is going on here, but
with this patch I could not reproduce the ACCESS_DENIED anymore. It might be
coincidence and this patch does not fix anything, but for me it does not really
hurt either.

Metze, please check!

Volker